+++ test/integration/telemetry_test.exs
@@ -18,32 +18,50 @@ defmodule Oban.Integration.TelemetryTest do
 
     {:ok, _} = start_supervised({Oban, @oban_opts})
 
-    %Job{id: succ_id} = insert_job!(%{ref: 1, action: "OK"})
-    %Job{id: fail_id} = insert_job!(%{ref: 2, action: "FAIL"})
+    %Job{id: success_id} = insert_job!(%{ref: 1, action: "OK"})
+    %Job{id: exception_id} = insert_job!(%{ref: 2, action: "FAIL"})
+    %Job{id: error_id} = insert_job!(%{ref: 2, action: "ERROR"})
 
-    assert_receive {:executed, :success, succ_duration, succ_meta}
-    assert_receive {:executed, :failure, fail_duration, fail_meta}
+    assert_receive {:executed, :success, success_duration, success_meta}
+    assert_receive {:executed, :failure, exception_duration, %{kind: :exception} = exception_meta}
+    assert_receive {:executed, :failure, error_duration, %{kind: :error} = error_meta}
 
-    assert succ_duration > 0
-    assert fail_duration > 0
+    assert success_duration > 0
+    assert exception_duration > 0
+    assert error_duration > 0
 
     assert %{
-             id: ^succ_id,
+             id: ^success_id,
              args: %{},
              queue: "zeta",
              worker: "Oban.Integration.Worker",
              attempt: 1,
              max_attempts: 20
-           } = succ_meta
+           } = success_meta
 
     assert %{
-             id: ^fail_id,
+             id: ^exception_id,
              args: %{},
              queue: "zeta",
              worker: "Oban.Integration.Worker",
              attempt: 1,
-             max_attempts: 20
-           } = fail_meta
+             max_attempts: 20,
+             kind: :exception,
+             error: _,
+             stack: [_ | _]
+           } = exception_meta
+
+    assert %{
+             id: ^error_id,
+             args: %{},
+             queue: "zeta",
+             worker: "Oban.Integration.Worker",
+             attempt: 1,
+             max_attempts: 20,
+             kind: :error,
+             error: "ERROR",
+             stack: [_ | _]
+           } = error_meta
 
     :ok = stop_supervised(Oban)
   end